feat: add base component for StudioContentMenu (#14090)
